fix: metadata stuff
