+++ apps/new-frontend/src/app/admin/page.tsx
@@ -1,5 +1,10 @@
 import { getUser, RoleEnum } from "@lib/auth";
 import { forbidden, unauthorized } from "next/navigation";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+	title: "Administrator Panel - Holy Grail",
+};
 
 export default async function AdminPage() {
 	const user = await getUser();
