+++ apps/new-frontend/src/app/developer/page.tsx
@@ -1,7 +1,12 @@
 import { getUser, RoleEnum } from "@lib/auth";
 import { forbidden, unauthorized } from "next/navigation";
+import type { Metadata } from "next";
 
-export default async function AdminPage() {
+export const metadata: Metadata = {
+	title: "Developer Panel - Holy Grail",
+};
+
+export default async function DeveloperPage() {
 	const user = await getUser();
 	if (!user) {
 		unauthorized();
